Flask/requirements.txt
flask
flask-cors
python-dotenv
sentence-transformers
langchain
langchain-chroma
langchain-huggingface
pypdf
langchain-google-genai
langchain-community
